+++ src/components/input-stuff.js
@@ -1,7 +1,5 @@
 import React from 'react';
-import DayPickerInput from "react-day-picker/DayPickerInput";
 
-import "react-day-picker/lib/style.css"
 
 const FancyInput = (props) => {
   return (
@@ -30,29 +28,6 @@ const FancyButton = (props) => {
     )
 }
 
-class FancyDatePicker extends React.Component {
-  handleBirthdayChange = day => {
-    this.setState({
-      birthday: day
-    })
-  }
-
-  render() {
-    return (
-      <form>
-        <DayPickerInput
-          name="birthday"
-          placeholder="DD/MM/YYYY"
-          format="DD/MM/YYYY"
-          onDayChange={ this.handleBirthdayChange }
-          dayPickerProps={{
-            enableOutsideDays: true,
-          }}
-        />
-      </form>
-    )
-  }
-}
 
 
-export {FancyInput, FancyDropdown, FancyButton, FancyDatePicker};
+export {FancyInput, FancyDropdown, FancyButton};
